Fix another round of VC++ warnings about mismatching signs and type conversions.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6340 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/Entity.cpp
@@ -2091,7 +2091,7 @@ void idEntity::Save( idSaveGame *savefile ) const
 
 	savefile->WriteInt(team);
 
-	savefile->WriteInt(m_EntityRelations.size());
+    savefile->WriteInt(static_cast<int>(m_EntityRelations.size()));
 	for (EntityRelationsMap::const_iterator i = m_EntityRelations.begin(); i != m_EntityRelations.end(); ++i)
 	{
 		savefile->WriteObject(i->first);
@@ -2158,7 +2158,7 @@ void idEntity::Save( idSaveGame *savefile ) const
 	for( int i=0; i < m_Attachments.Num(); i++ )
 		m_Attachments[i].Save( savefile );
 
-	savefile->WriteInt( m_AttNameMap.size() );
+    savefile->WriteInt(static_cast<int>(m_AttNameMap.size()));
 	for ( AttNameMap::const_iterator k = m_AttNameMap.begin();
          k != m_AttNameMap.end(); ++k )
     {
@@ -2210,7 +2210,7 @@ void idEntity::Save( idSaveGame *savefile ) const
 	savefile->WriteVec3( m_VinePlantNormal );
 
 	// SteveL #3817: make decals persistent
-	savefile->WriteInt( decals_list.size() );
+    savefile->WriteInt(static_cast<int>(decals_list.size()));
 	for ( std::list<SDecalInfo>::const_iterator i = decals_list.begin(); i != decals_list.end(); ++i )
 	{
 		savefile->WriteString( i->decal );
@@ -6922,7 +6922,7 @@ bool idEntity::HandleGuiCommands( idEntity *entityGui, const char *cmds ) {
 	if ( entityGui && cmds && *cmds ) {
 		idLexer src;
 		idToken token, token2, token3, token4;
-		src.LoadMemory( cmds, strlen( cmds ), "guiCommands" );
+        src.LoadMemory(cmds, static_cast<int>(strlen(cmds)), "guiCommands");
 		while( 1 ) {
 
 			if ( !src.ReadToken( &token ) ) {
@@ -7284,7 +7284,7 @@ idCurve_Spline<idVec3> *idEntity::GetSpline( void ) const {
 		return NULL;
 	}
 
-	idStr str = kv->GetKey().Right( kv->GetKey().Length() - strlen( curveTag ) );
+    idStr str = kv->GetKey().Right(kv->GetKey().Length() - static_cast<int>(strlen(curveTag)));
 	if ( str.Icmp( "CatmullRomSpline" ) == 0 ) {
 		spline = new idCurve_CatmullRomSpline<idVec3>();
 	} else if ( str.Icmp( "nubs" ) == 0 ) {
